CS 9.8 release

+++ SRC/systemVariables.cpp
@@ -296,7 +296,7 @@ static char* Szulutime(char* value)
     GetTimeInfo(&ptm,true,true);
     uint64 elapsedMS = ElapsedMilliseconds();
 	sprintf(systemValue,(char*)"%d-%2.2d-%2.2dT%2.2d:%2.2d:%2.2d.%3.3dZ",ptm.tm_year+1900,ptm.tm_mon+1,ptm.tm_mday,
-		ptm.tm_hour,ptm.tm_min,ptm.tm_sec, elapsedMS%1000);
+		ptm.tm_hour,ptm.tm_min,ptm.tm_sec, (int)(elapsedMS%1000));
     return systemValue;
 }
